Fixed compilation warning

+++ drivers/net/e1000.c
@@ -5015,7 +5015,7 @@ static int e1000_transmit(struct eth_device *nic, void *packet, int length)
 	txp = tx_base + tx_tail;
 	tx_tail = (tx_tail + 1) % 8;
 
-	txp->buffer_addr = cpu_to_le64(virt_to_bus(hw->pdev, nv_packet));
+	txp->buffer_addr = (uint64_t)(cpu_to_le64(virt_to_bus(hw->pdev, nv_packet)));
 	txp->lower.data = cpu_to_le32(hw->txd_cmd | length);
 	txp->upper.data = 0;
 	E1000_WRITE_REG(hw, TDT, tx_tail);
